Split get_batch_size
