+++ poutyne/__init__.py
@@ -24,3 +24,4 @@ from .utils import *
 from .plotting import *
 from .framework import *
 from .layers import *
+from .warning_manager import *
