+++ poutyne/framework/__init__.py
@@ -23,4 +23,3 @@ from .callbacks import *
 from .metrics import *
 from .experiment import *
 from .model_bundle import *
-from .warning_manager import *
